Abstract filterMoviesByGenre() to it's own file
