+++ src/components/filterMoviesByGenre.js
@@ -1,0 +1,13 @@
+const filterMoviesByGenre = (movies, genre) =>
+{
+    let genreInt = parseInt(genre);
+
+    let filteredResults = movies.filter(function(movie)
+    {
+        return movie.genre_ids.indexOf(genreInt) > -1;
+    });
+
+    return filteredResults;
+}
+
+export default filterMoviesByGenre;
